# Set types as attributes on the cstruct object