… tests/test_annotations.py
… tests/test_basic.py
… tests/test_parser.py
… tests/test_tools_stubgen.py